Move labels out of the helper package 🌿
